suppression warning

+++ Moteur/Source/IndexBuffer.cpp
@@ -1,8 +1,8 @@
 #include "IndexBuffer.h"
 
-IndexBuffer::IndexBuffer(std::vector<UINT16> m_indices) : Buffer(m_indices.size() * sizeof(UINT16), m_indices.data())
+IndexBuffer::IndexBuffer(std::vector<UINT16> m_indices) : Buffer((UINT)(m_indices.size() * sizeof(UINT16)), m_indices.data())
 {
-	m_nbVertex = m_indices.size();
+	m_nbVertex = (UINT)m_indices.size();
 
 	// Création de la vue de l'index buffer
 	m_indexBufferView.BufferLocation = m_buffer->GetGPUVirtualAddress();
